+++ WeatherApp/WebMvc/ViewModels/WeatherExtremesViewModel.cs
@@ -7,7 +7,7 @@ namespace WebMvc.ViewModels
     {
         public LocationModel Location { get; set; }
         public DateTime LastUpdate { get; set; }
-        public int MinimumTemperature { get; set; }
-        public int MaximumWindSpeed { get; set; }
+        public double MinimumTemperature { get; set; }
+        public double MaximumWindSpeed { get; set; }
     }
 }
